[MISC] 공용 컴포넌트화
